feat: add bom publication
